application setup

+++ engine/src/yuki/core/memory/memory_module.h
@@ -25,7 +25,7 @@ yuki_memory_tag;
 
 YUKI_API bool
 memory_module_startup
-(u32 *required_memory_size, void *state);
+(u64 *required_memory_size, void *state);
 
 YUKI_API bool
 memory_module_shutdown
@@ -34,24 +34,24 @@ memory_module_shutdown
 
 YUKI_API void *
 memory_module_allocate_block
-(const u32 block_size, const yuki_memory_tag tag);
+(const u64 block_size, const yuki_memory_tag tag);
 
 YUKI_API void
 memory_module_deallocate_block
-(void *memory_block, const u32 block_size, const yuki_memory_tag tag);
+(void *memory_block, const u64 block_size, const yuki_memory_tag tag);
 
 
 YUKI_API void *
 memory_module_set_block_zero
-(void *memory_block, const u32 block_size);
+(void *memory_block, const u64 block_size);
 
 YUKI_API void *
 memory_module_set_block_value
-(void *memory_block, const u32 block_size, const s32 value);
+(void *memory_block, const u64 block_size, const s32 value);
 
 YUKI_API void *
 memory_module_copy_block
-(void *dst_block, void *src_block, const u32 block_size);
+(void *dst_block, void *src_block, const u64 block_size);
 
 
 YUKI_API const_str
